Clippy & Remove old markdown example
