+++ justfile
@@ -1,11 +1,11 @@
 build-markdown:
-	cd examples/markdown && trunk build --example markdown2
+	cd examples/markdown && trunk build --example markdown
 
 dist-out := "/tmp/massive-shell-dist"
 
 build-markdown-release:
 	rm -f shell/dist/*
-	cd examples/markdown && trunk build --example markdown2 --release
+	cd examples/markdown && trunk build --example markdown --release
 	mkdir -p {{dist-out}}
 	rm -f {{dist-out}}/*
 	cp examples/markdown/dist/massive-shell-*.js {{dist-out}}/massive-markdown.js
@@ -22,7 +22,7 @@ build-code-viewer-release:
 	sed -i '' 's/massive-shell_bg.wasm/massive-code_bg.wasm/g' {{dist-out}}/massive-code.js
 
 serve-markdown:
-	cd examples/markdown && trunk serve --example markdown2 --port 8888 --no-minification
+	cd examples/markdown && trunk serve --example markdown --port 8888 --no-minification
 
 serve-markdown-release:
 	cd examples/markdown && trunk serve --example markdown --port 8888 --no-minification --release
